requirements.txt
maturin
python-dateutil
polars
hypothesis
numpy
pandas
pytest
holidays
mypy
black
ruff
